# C# performance: further reduce allocations by pooling OutPt, OutRec, reusing Active and tracking OutPt count in outRec