--- a/CSharp/Clipper2Lib.Benchmark/Program.cs
+++ b/CSharp/Clipper2Lib.Benchmark/Program.cs
@@ -1,8 +1,12 @@
 using BenchmarkDotNet.Running;
 
+#if USINGZ
+namespace Clipper2ZLib.Benchmark
+#else
 namespace Clipper2Lib.Benchmark
+#endif
 {
-    public static class Program
+  public static class Program
     {        
         public static void Main()
         {