--- a/CSharp/Clipper2Lib.Examples/ConsoleDemo/Main.cs
+++ b/CSharp/Clipper2Lib.Examples/ConsoleDemo/Main.cs
@@ -8,7 +8,11 @@
 
 using System.Reflection;
 using System.IO;
+#if USINGZ
+using Clipper2ZLib;
+#else
 using Clipper2Lib;
+#endif
 using System;
 
 namespace ClipperDemo1