--- a/CSharp/Clipper2Lib/Clipper.Minkowski.cs
+++ b/CSharp/Clipper2Lib/Clipper.Minkowski.cs
@@ -10,7 +10,11 @@
 #nullable enable
 using System;
 
+#if USINGZ
+namespace Clipper2ZLib
+#else
 namespace Clipper2Lib
+#endif
 {
   public static class Minkowski
   {