--- a/CSharp/Clipper2Lib/HashCode.cs
+++ b/CSharp/Clipper2Lib/HashCode.cs
@@ -3,7 +3,11 @@ using System.ComponentModel;
 using System.Runtime.CompilerServices;
 using System.Security.Cryptography;
 
+#if USINGZ
+namespace Clipper2ZLib
+#else
 namespace Clipper2Lib
+#endif
 {
   /*