I have analyzed and reduced C# allocations done by clipper2.
This pull request finalizes the work [in my last one](https://github.com/AngusJohnson/Clipper2/pull/1010#issue-3425529123) and contains the following optimizations:

1. track the number of OutPt per OutRec to allocate Path64 with correct capacity
2. reuse HorzJoin objects by using new derived class HorzJoinPoolList
3. reuse OutRec objects by using new derived class OutRecPoolList
4. round up requested capacity to next power of 2 for all classes derived from abstract PooledList<T>
5. eliminate boxing for PooledList<T>.GetEnumerator()
6. reuse OutPt objects by using new derived class OutPtPoolList
7. reuse Active objects by putting them on a free stack

I have benchmarked and tested these changes. The summary compared to current main and last release 1.5.4:

| Benchmark      | Mean     | Error   | StdDev  | Allocated |
|---------------:|---------:|--------:|--------:|----------:|
| Baseline_1.5.4 | 232.9 ms | 1.55 ms | 1.21 ms | 278.66 MB |
| Baseline_Main  | 223.5 ms | 1.35 ms | 1.05 ms | 223.87 MB |
| Result         | 220.0 ms | 1.39 ms | 1.16 ms | 159.36 MB |

While my previous PR cut allocation volume to 80%, this one further improves down to 57% when Clipper2 objects are cleared and reused. In this single thread benchmark the runtime difference is a minor 1%. More importantly it is never a downgrade. But I also want to provide benchmarks where the difference is more significant. The following benchmarks use a mutli-threaded workload and the .NET default garbage collector mode "workstation mode". This scenario puts significant pressure on the GC because there are more threads allocating than collecting garbage. This results in about 10% faster execution, 25% compared to v1.5.4.

### Baseline 1: Clipper2 1.5.4
| Method             | Mean       | Error    | StdDev   | Gen0        | Completed Work Items | Lock Contentions | Gen1       | Allocated  |
|------------------- |-----------:|---------:|---------:|------------:|---------------------:|-----------------:|-----------:|-----------:|
| Multi-Thread 1     |   109.6 ms |  2.17 ms |  5.32 ms |  18000.0000 |             697.0000 |                - |  1000.0000 |     901 MB |
| Multi-Thread 2     |   114.6 ms |  2.25 ms |  5.02 ms |  19000.0000 |             838.0000 |           6.0000 |  1000.0000 |  926.31 MB |
| Multi-Thread 3     | 1,092.7 ms | 21.83 ms | 30.60 ms | 116000.0000 |             753.0000 |          56.0000 |  9000.0000 | 5502.83 MB |
| Multi-Thread 4     | 1,096.8 ms | 21.72 ms | 41.84 ms | 118000.0000 |             839.0000 |          47.0000 |  9000.0000 | 5600.38 MB |

### Baseline 2: Clipper2 main
| Method             | Mean      | Error     | StdDev    | Gen0       | Completed Work Items | Lock Contentions | Gen1       | Allocated  |
|------------------- |----------:|----------:|----------:|-----------:|---------------------:|-----------------:|-----------:|-----------:|
| Multi-Thread 1     |  92.08 ms |  1.835 ms |  4.431 ms | 12000.0000 |             670.0000 |           3.0000 |  1000.0000 |  615.94 MB |
| Multi-Thread 2     |  97.76 ms |  1.939 ms |  4.376 ms | 12000.0000 |             636.0000 |                - |  1000.0000 |  607.54 MB |
| Multi-Thread 3     | 922.70 ms | 18.425 ms | 32.750 ms | 81000.0000 |             753.0000 |          36.0000 | 10000.0000 | 3811.72 MB |
| Multi-Thread 4     | 935.23 ms | 18.288 ms | 33.441 ms | 83000.0000 |             850.0000 |          36.0000 | 11000.0000 |  3918.8 MB |

### Result
| Method             | Mean      | Error     | StdDev    | Gen0       | Completed Work Items | Lock Contentions | Gen1       | Allocated  |
|------------------- |----------:|----------:|----------:|-----------:|---------------------:|-----------------:|-----------:|-----------:|
| Multi-Thread 1     |  82.16 ms |  1.640 ms |  3.312 ms |  8000.0000 |             507.0000 |                - |  1000.0000 |  413.68 MB |
| Multi-Thread 2     |  86.51 ms |  1.708 ms |  2.758 ms |  9000.0000 |             829.0000 |                - |  1000.0000 |  443.79 MB |
| Multi-Thread 3     | 810.19 ms | 15.744 ms | 18.131 ms | 51000.0000 |             984.0000 |          33.0000 | 11000.0000 |  2416.9 MB |
| Multi-Thread 4     | 891.52 ms | 18.042 ms | 52.915 ms | 51000.0000 |             755.0000 |          42.0000 | 12000.0000 | 2424.26 MB |

This is everything to know about the results, below I will provide my benchmark findings and reasoning. It is quite a lot and an optional read.

### Reasoning for implementation choices
For change 1-5 the reasoning is clear. Lists for HorzJoin and OutRec already existed and are replaced with PooledList implementations like Vertex. I have not done that for HorzSegment because it would have required to port Sort() methods to PooledList and HorzSegment objects are allocated infrequently.
OutPt and Active are the major offenders regarding allocations. For OutPt I added a new pooled list to also use arena-style allocation, because this simplifies the changes. The exception is Active, which showed significantly different behaviour in benchmarks. Actives are large objects (~160 bytes, 3 cache lines) and accessed frequently. Pooling them arena allocation style caused performance to drop significantly due to CPU cache and prefetch effects. Allocating in C# is very predictable for the CPU (only a pointer increment) and has good spatial locality. So I changed from an arena to a stack based reuse when Actives are deleted from the AEL. This keeps the allocated working set smaller and more likely in L1 cache. Also note this will reduce allocations even if clipper static methods are used.
I additionally benchmarked to clear this free stack in ClearSolutionOnly. This has shown to be the overall best choice because the cache misses from chasing pointers to reused Actives seem to have a major impact after the CPU did something else than Clipper operations. Detailed benchmark results for these findings are below for reference.

### Single Thread benchmark with all changes
This detailed single thread benchmark breaks down the impact on runtime allocations of each change. OutPt is most significant. Active is also allocating a lot, but always recycling (not clearing the free stack in ClearSolutionOnly) gives a perfromance hit altough it reduces allocations. Notice the additional options and branches in this breakdown make it overall slower than the final result.
| UseOutPtPool | UseHorzJoinPool | UseOutRecPool | TrackOutRecPathCapacity | ActiveAllocMode | Mean     | Error   | StdDev  | Allocated |
|------------- |---------------- |-------------- |------------------------ |---------------- |---------:|--------:|--------:|----------:|
| False        | False           | False         | False                   | ALWAYS_ALLOC    | 239.2 ms | 2.21 ms | 1.85 ms | 227.34 MB |
| False        | False           | False         | False                   | ALWAYS_RECYCLE  | 241.9 ms | 3.55 ms | 2.97 ms | 185.08 MB |
| False        | False           | False         | False                   | FREE_ON_CLEAR   | 228.8 ms | 3.22 ms | 2.51 ms | 204.26 MB |
| False        | False           | False         | True                    | ALWAYS_ALLOC    | 230.5 ms | 4.48 ms | 4.19 ms | 214.66 MB |
| False        | False           | False         | True                    | ALWAYS_RECYCLE  | 237.2 ms | 3.58 ms | 2.80 ms | 172.41 MB |
| False        | False           | False         | True                    | FREE_ON_CLEAR   | 228.2 ms | 2.14 ms | 2.10 ms | 191.59 MB |
| False        | False           | True          | False                   | ALWAYS_ALLOC    | 233.8 ms | 4.30 ms | 3.35 ms | 220.58 MB |
| False        | False           | True          | False                   | ALWAYS_RECYCLE  | 236.3 ms | 2.48 ms | 2.07 ms | 178.33 MB |
| False        | False           | True          | False                   | FREE_ON_CLEAR   | 228.0 ms | 2.96 ms | 2.31 ms | 197.51 MB |
| False        | False           | True          | True                    | ALWAYS_ALLOC    | 231.6 ms | 2.89 ms | 2.41 ms | 207.91 MB |
| False        | False           | True          | True                    | ALWAYS_RECYCLE  | 238.3 ms | 3.88 ms | 3.03 ms | 165.65 MB |
| False        | False           | True          | True                    | FREE_ON_CLEAR   | 232.2 ms | 3.84 ms | 3.00 ms | 184.83 MB |
| False        | True            | False         | False                   | ALWAYS_ALLOC    | 232.2 ms | 3.48 ms | 2.90 ms | 227.31 MB |
| False        | True            | False         | False                   | ALWAYS_RECYCLE  | 238.7 ms | 2.58 ms | 2.01 ms | 185.06 MB |
| False        | True            | False         | False                   | FREE_ON_CLEAR   | 226.4 ms | 2.37 ms | 1.98 ms | 204.24 MB |
| False        | True            | False         | True                    | ALWAYS_ALLOC    | 231.5 ms | 4.08 ms | 3.41 ms | 214.64 MB |
| False        | True            | False         | True                    | ALWAYS_RECYCLE  | 239.9 ms | 2.12 ms | 1.77 ms | 172.38 MB |
| False        | True            | False         | True                    | FREE_ON_CLEAR   | 224.3 ms | 1.20 ms | 0.94 ms | 191.56 MB |
| False        | True            | True          | False                   | ALWAYS_ALLOC    | 232.2 ms | 2.30 ms | 2.26 ms | 220.56 MB |
| False        | True            | True          | False                   | ALWAYS_RECYCLE  | 237.9 ms | 3.16 ms | 2.64 ms |  178.3 MB |
| False        | True            | True          | False                   | FREE_ON_CLEAR   | 227.1 ms | 2.17 ms | 1.81 ms | 197.48 MB |
| False        | True            | True          | True                    | ALWAYS_ALLOC    | 229.1 ms | 2.87 ms | 2.24 ms | 207.88 MB |
| False        | True            | True          | True                    | ALWAYS_RECYCLE  | 234.0 ms | 2.68 ms | 2.09 ms | 165.63 MB |
| False        | True            | True          | True                    | FREE_ON_CLEAR   | 228.2 ms | 2.83 ms | 2.21 ms | 184.81 MB |
| True         | False           | False         | False                   | ALWAYS_ALLOC    | 228.3 ms | 2.59 ms | 2.02 ms | 201.89 MB |
| True         | False           | False         | False                   | ALWAYS_RECYCLE  | 243.1 ms | 2.00 ms | 1.67 ms | 159.64 MB |
| True         | False           | False         | False                   | FREE_ON_CLEAR   | 227.3 ms | 2.37 ms | 1.85 ms | 178.82 MB |
| True         | False           | False         | True                    | ALWAYS_ALLOC    | 226.5 ms | 2.40 ms | 1.87 ms | 189.22 MB |
| True         | False           | False         | True                    | ALWAYS_RECYCLE  | 231.0 ms | 2.81 ms | 2.19 ms | 146.97 MB |
| True         | False           | False         | True                    | FREE_ON_CLEAR   | 223.6 ms | 2.98 ms | 2.33 ms | 166.15 MB |
| True         | False           | True          | False                   | ALWAYS_ALLOC    | 230.9 ms | 3.96 ms | 3.31 ms | 195.14 MB |
| True         | False           | True          | False                   | ALWAYS_RECYCLE  | 230.9 ms | 2.06 ms | 1.72 ms | 152.88 MB |
| True         | False           | True          | False                   | FREE_ON_CLEAR   | 234.3 ms | 2.30 ms | 1.92 ms | 172.06 MB |
| True         | False           | True          | True                    | ALWAYS_ALLOC    | 222.7 ms | 1.90 ms | 1.59 ms | 182.46 MB |
| True         | False           | True          | True                    | ALWAYS_RECYCLE  | 228.5 ms | 1.96 ms | 1.53 ms | 140.21 MB |
| True         | False           | True          | True                    | FREE_ON_CLEAR   | 221.5 ms | 1.16 ms | 0.96 ms | 159.39 MB |
| True         | True            | False         | False                   | ALWAYS_ALLOC    | 228.2 ms | 2.75 ms | 2.30 ms | 201.87 MB |
| True         | True            | False         | False                   | ALWAYS_RECYCLE  | 229.9 ms | 1.85 ms | 1.55 ms | 159.61 MB |
| True         | True            | False         | False                   | FREE_ON_CLEAR   | 224.2 ms | 1.83 ms | 1.53 ms |  178.8 MB |
| True         | True            | False         | True                    | ALWAYS_ALLOC    | 224.6 ms | 1.63 ms | 1.36 ms | 189.19 MB |
| True         | True            | False         | True                    | ALWAYS_RECYCLE  | 227.8 ms | 2.81 ms | 2.35 ms | 146.94 MB |
| True         | True            | False         | True                    | FREE_ON_CLEAR   | 219.5 ms | 1.56 ms | 1.30 ms | 166.12 MB |
| True         | True            | True          | False                   | ALWAYS_ALLOC    | 223.9 ms | 2.25 ms | 1.75 ms | 195.11 MB |
| True         | True            | True          | False                   | ALWAYS_RECYCLE  | 236.4 ms | 2.23 ms | 1.86 ms | 152.86 MB |
| True         | True            | True          | False                   | FREE_ON_CLEAR   | 223.0 ms | 2.80 ms | 2.34 ms | 172.04 MB |
| True         | True            | True          | True                    | ALWAYS_ALLOC    | 223.5 ms | 3.29 ms | 2.74 ms | 182.44 MB |
| True         | True            | True          | True                    | ALWAYS_RECYCLE  | 225.1 ms | 1.48 ms | 1.15 ms | 140.18 MB |
| True         | True            | True          | True                    | FREE_ON_CLEAR   | 220.9 ms | 2.77 ms | 2.16 ms | 159.36 MB |

### Multi-Thread benchmark server GC
The Multi-Thread benchmarks in server garbage collection mode show performance drops when always recycling Active. The measurements have larger standard deviation than single threaded because of thread coordination.
| Method             | ActiveAllocMode | Mean      | Error     | StdDev    | Median    | Completed Work Items | Lock Contentions | Gen0      | Gen1      | Allocated  |
|------------------- |---------------- |----------:|----------:|----------:|----------:|---------------------:|-----------------:|----------:|----------:|-----------:|
| Multi-Thread 1     | ALWAYS_ALLOC    |  83.80 ms |  6.525 ms | 18.721 ms |  76.25 ms |             617.0000 |           1.0000 |         - |         - |  439.08 MB |
| Multi-Thread 1     | ALWAYS_RECYCLE  |  82.21 ms |  4.662 ms | 12.917 ms |  78.73 ms |             918.0000 |           1.0000 |         - |         - |  328.89 MB |
| Multi-Thread 1     | FREE_ON_CLEAR   |  98.67 ms | 12.288 ms | 36.231 ms |  81.47 ms |             752.0000 |                - |         - |         - |  409.31 MB |
| Multi-Thread 2     | ALWAYS_ALLOC    |  79.14 ms |  1.577 ms |  4.043 ms |  78.25 ms |             905.0000 |           2.0000 |         - |         - |  495.21 MB |
| Multi-Thread 2     | ALWAYS_RECYCLE  |  82.37 ms |  2.009 ms |  5.601 ms |  81.51 ms |             731.0000 |                - |         - |         - |   343.7 MB |
| Multi-Thread 2     | FREE_ON_CLEAR   |  77.86 ms |  1.796 ms |  4.856 ms |  76.76 ms |             834.0000 |                - |         - |         - |  432.15 MB |
| Multi-Thread 3     | ALWAYS_ALLOC    | 651.09 ms | 12.985 ms | 34.660 ms | 651.26 ms |            1021.0000 |          60.0000 | 1000.0000 |         - | 2600.15 MB |
| Multi-Thread 3     | ALWAYS_RECYCLE  | 735.79 ms | 14.506 ms | 34.193 ms | 737.05 ms |            1021.0000 |          50.0000 | 1000.0000 |         - | 1834.96 MB |
| Multi-Thread 3     | FREE_ON_CLEAR   | 650.82 ms | 13.880 ms | 40.490 ms | 646.63 ms |             879.0000 |          40.0000 | 1000.0000 |         - | 2372.08 MB |
| Multi-Thread 4     | ALWAYS_ALLOC    | 671.55 ms | 13.382 ms | 36.633 ms | 675.01 ms |             809.0000 |          50.0000 | 1000.0000 |         - |  2606.4 MB |
| Multi-Thread 4     | ALWAYS_RECYCLE  | 739.99 ms | 14.664 ms | 37.324 ms | 734.48 ms |            1139.0000 |          43.0000 | 1000.0000 |         - | 1856.66 MB |
| Multi-Thread 4     | FREE_ON_CLEAR   | 655.10 ms | 12.981 ms | 36.826 ms | 652.43 ms |            1023.0000 |          56.0000 | 1000.0000 |         - | 2400.58 MB |


### benchmark hardware
BenchmarkDotNet v0.15.2, Windows 11 (10.0.26100.6584/24H2/2024Update/HudsonValley)
AMD Ryzen 9 7950X3D 4.20GHz, 1 CPU, 32 logical and 16 physical cores
.NET SDK 9.0.304
  [Host]     : .NET 8.0.19 (8.0.1925.36514), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
workstation or server GC as stated above each table